Feat: adjustments to OG kickstart.nvim config
