+++ lua/custom/plugins/treesitter.lua
@@ -1,0 +1,74 @@
+--=============================================================================
+-- SYNTAX + PARSING (nvim-treesitter)
+
+-- Better syntax highlighting/indenting using parsers instead of regex (more accurate + feature-rich).
+--=============================================================================
+
+return { -- Highlight, edit, and navigate code
+  'nvim-treesitter/nvim-treesitter',
+  build = ':TSUpdate',
+  event = { 'BufReadPost', 'BufNewFile' },
+  main = 'nvim-treesitter.configs', -- Sets main module to use for opts
+
+  -- [[ Configure Treesitter ]] See `:help nvim-treesitter`
+  opts = {
+    ensure_installed = {
+      'bash',
+      'c',
+      'diff',
+      'html',
+      'css',
+      'javascript',
+      'php',
+      'lua',
+      'luadoc',
+      'markdown',
+      'markdown_inline',
+      'query',
+      'vim',
+      'vimdoc',
+    },
+    -- Autoinstall languages that are not installed
+    auto_install = false,
+    highlight = {
+      enable = true,
+      -- Some languages depend on vim's regex highlighting system (such as Ruby) for indent rules.
+      --  If you are experiencing weird indenting issues, add the language to
+      --  the list of additional_vim_regex_highlighting and disabled languages for indent.
+      additional_vim_regex_highlighting = { 'ruby' },
+    },
+    indent = { enable = true, disable = { 'ruby' } },
+  },
+
+  -- There are additional nvim-treesitter modules that you can use to interact
+  -- with nvim-treesitter. You should go explore a few and see what interests you:
+  --
+  --    - Incremental selection: Included, see `:help nvim-treesitter-incremental-selection-mod`
+  --    - Show your current context: https://github.com/nvim-treesitter/nvim-treesitter-context
+  --    - Treesitter + textobjects: https://github.com/nvim-treesitter/nvim-treesitter-textobjects
+
+  -- Blade-specific extra config
+  config = function(_, opts)
+    -- Keep the normal Treesitter setup using your opts
+    require('nvim-treesitter.configs').setup(opts)
+
+    -- Blade Treesitter config (Laravel Blade templates)
+    local parser_config = require('nvim-treesitter.parsers').get_parser_configs()
+
+    parser_config.blade = {
+      install_info = {
+        url = 'https://github.com/EmranMR/tree-sitter-blade',
+        files = { 'src/parser.c' },
+        branch = 'main',
+      },
+      filetype = 'blade',
+    }
+
+    -- Tell Neovim that *.blade.php files are "blade" filetype
+    vim.filetype.add {
+      pattern = {
+        ['.*%.blade%.php'] = 'blade',
+      },
+    }
+  end,
+}
